Poll Details
Working on the Poll component that will show the poll details.  I have
currently commented out the list html in the Dashboard component.  Once
the Poll details is complete, The list items will have a button that
sends the user to the details.  Currently working on the output of
answered questions