Changed Login Component
Changed the Login component to make all the user icons as buttons
instead of links.  Added Routes to the App component.  Removed the
history.push('/') so when a user logs in, they will be able to see
whatever page they were on.  This allows the testing of seeing the
page /question/:id if the id does not exist, there will be a page that
shows up that says the question does not exist